feat(controller): update business logic in task controller

+++ src/AppBundle/Controller/TaskController.php
@@ -4,6 +4,7 @@ namespace AppBundle\Controller;
 
 use AppBundle\Entity\Task;
 use AppBundle\Form\TaskType;
+use AppBundle\Security\TaskVoter;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
@@ -15,9 +16,15 @@ class TaskController extends Controller
      */
     public function listAction()
     {
-        $tasksList = $this->getDoctrine()->getRepository('AppBundle:Task')->findBy(
-            ['user' => $this->getUser()]
-        );
+        if ($this->getUser()->hasRole('ROLE_ADMIN')) {
+            $tasksList = $this->getDoctrine()->getRepository('AppBundle:Task')->findUserAndAnonymousTask(
+                $this->getUser()
+            );
+        } else {
+            $tasksList = $this->getDoctrine()->getRepository('AppBundle:Task')->findBy(
+                ['user' => $this->getUser()]
+            );
+        }
 
         return $this->render('task/list.html.twig', ['tasks' => $tasksList]);
     }
@@ -53,6 +60,8 @@ class TaskController extends Controller
      */
     public function editAction(Task $task, Request $request)
     {
+        $this->denyAccessUnlessGranted(TaskVoter::EDIT, $task);
+
         $form = $this->createForm(TaskType::class, $task);
 
         $form->handleRequest($request);
@@ -76,6 +85,8 @@ class TaskController extends Controller
      */
     public function toggleTaskAction(Task $task)
     {
+        $this->denyAccessUnlessGranted(TaskVoter::EDIT, $task);
+
         $task->toggle(!$task->isDone());
         $this->getDoctrine()->getManager()->flush();
 
@@ -89,6 +100,8 @@ class TaskController extends Controller
      */
     public function deleteTaskAction(Task $task)
     {
+        $this->denyAccessUnlessGranted(TaskVoter::DELETE, $task);
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($task);
         $em->flush();
